testing out sparse mask
